dom/url/tests/test_unknown_url_origin.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for unknwon URL.origin</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript">

  is((new URL("blob:http://foo.com/bar")).origin, "http://foo.com");
  is((new URL("blob:blob:http://foo.com/bar")).origin, "http://foo.com");

  </script>
</body>
</html>
